Agregando funcionalidad a editar productos

+++ scripts/predefinidos.js
@@ -166,4 +166,8 @@ const productos = [
     }
 ];
 
-localStorage.setItem("-productos",JSON.stringify(productos));
+const existProduct = JSON.parse(localStorage.getItem("-productos")) || [];
+
+if(existProduct.length = 0){
+    localStorage.setItem("-productos",JSON.stringify(productos));
+}
